update and delete some not neccesary file
