--- computer-confugration-website/src/app/complete-build/complete_db.js
@@ -1,12 +1,0 @@
-const builds = [
-    {
-        id: 1,
-        name: "Exceptional Black Build with Core i7-12700F",
-        describe_1: "Intel Core i7-12700F | GeForce RTX 3060 12GB",
-        describe_2: "16GB DDR4 3200 | 500GB SSD",
-        describe_3: "LGA170 Semi-Modular 600W",
-        price: 1500,
-    }
-];
-
-export default builds;
